Add version update workflow (#4)
Fix the workflow for making automated version update PRs and make it callable. Previously, the workflow never made a PR because the `tee` command had an exit code of 0, so it always thought the package version was in sync even if it wasn't. This workflow first checks the exit status of `npm version` before deciding whether or not to make a PR. Also, we don't want a tag to be automatically tied to the commit that's made for the version update, so the `--no-git-tag-version` flag was added and the package file changes are manually added and committed.

J=SLAP-2004, SLAP-1931
TEST=manual

Add this workflow to a test branch in Headless. Call it from the SDK and check that if the package version is outdated, an automated PR is made to update it. If the package version matches the branch version, see that no PR is made.