Add workflow variables (#8)
Add variables to customize the node matrix in the `run_tests` workflow and the `fetch-depth` for the checkout action in the `percy_snapshots` and `wcag_test` workflows. Also, move the workflows out of the `.github` directory so they are not run on PRs in this repo.

J=SLAP-2005
TEST=manual

Test in the component lib and SDK that these variables can be used to customize the workflows as expected.