Move workflows (#9)
Move workflows back under the `.github` directory. Reusable workflows are required to be located directly under `.github/workflows` to be considered valid workflows. Since these workflows only run when called, these will not be arbitrarily run on commits or PRs in this repo.

J=SLAP-2005
TEST=manual

See that workflows fail when calling a reusable workflow not in `.github/workflows` and pass after reverting the change.