Update set-output calls (#25)
Update the calls to `set-output` since it is deprecated and will be removed. Followed the suggestion [here](https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/) and, for our `get-coverage-percent` action, upgraded to v1.10.0 of `@actions/core` which updates how it sets output variables under the hood. Also, updated the `Endbug/add-and-commit` action from v7 to v9, which uses the updated way to set output variables.

Note: There is still one `set-output` warning left, which comes from `coverallsapp/github-action` in the `coverage` workflow. There is already an [issue](https://github.com/coverallsapp/github-action/issues/135) made on that repo and it seems to be waiting on the merging of [this PR](https://github.com/coverallsapp/github-action/pull/136). Since we are pointing to their `master` branch, once the PR is merged, we will get the fix without any action needed on our part.

J=SLAP-2427
TEST=manual

Tested with a forked version of Core and ran tests on `search-ui-react` for the Percy and WCAG workflows. Saw that there were no more warnings for `set-output` (except for the one mentioned above).