Node 20 upgrade for get-coverage-percent (#33)
Update the node version of the get-coverage-percent action to 20.

J=WAT-3812
TEST=manual

Made a new branch on a repo which used the coverage workflow which uses this action, and temporarily updated the workflow and the branch on the other repo to reference this branch. [Triggered the workflow](https://github.com/yext/answers-search-ui/actions/runs/8395734213) and saw the warning about using a deprecated version of node was no longer present.