Merge main into v1 (#53)
* update verify-publish readme

* remove branch name from checkout step in publish workflow because some of our SDKs have main branch instead of master. Plus we should be able to publish non stable version from dev anyway

* Update run_tests workflow to work with multi-line scripts (#48)

Our vulnerabilities fixes cause this workflow to fail when the input is a multi-line script because environment variables are passed as single-line strings

TEST=manual
confirmed this workflow pass in a test PR in search-ui-react

* Update coverage workflow to work with multi-line scripts (#49)

Our vulnerabilities fixes cause this workflow to fail when the input is a multi-line script because environment variables are passed as single-line strings

TEST=manual
confirmed this workflow pass in a test PR in search-ui-react

* fix wcag test

* Fix run-script and verify-publish workflows (#52)

When trying to publish `search-ui-react`, I found that publishing a non latest (e.g beta) flag no longer works after my last refactoring attempt. Also found that the run-test action would fail if a working_directory was specified.

TEST=manual
successfully published my forked repo with beta tag.
saw github action pass then a working_directory is specified.